+++ ebayscraper/src/utils/postgres_utils.py
